moving tag replacement regex stuff to the relevant router
