+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/website/KifiSiteRouterTest.scala
@@ -163,26 +163,6 @@ class KifiSiteRouterTest extends Specification with ShoeboxTestInjector {
         status(result) must equalTo(404);
       }
     }
-
-    "substitute meta properties" in {
-      def substitute(property: String, newContent: String)(html: String) = {
-        val (pattern, newValue) = KifiSiteRouter.substituteMetaProperty(property, newContent)
-        pattern.replaceAllIn(html, newValue)
-      }
-      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge" />""") === """<meta property="og:title" content="Hi there!"/>"""
-      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge">""") === """<meta property="og:title" content="Hi there!"/>"""
-      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge" >""") === """<meta property="og:title" content="Hi there!"/>"""
-      substitute("og:description", "Hi there!")("""<meta property="og:description" content="Connecting People With Knowledge" />""") === """<meta property="og:description" content="Hi there!"/>"""
-      substitute("og:title", "Hi there!")("""<meta property="og:description" content="Connecting People With Knowledge" />""") === """<meta property="og:description" content="Connecting People With Knowledge" />"""
-    }
-
-    "substitute link tags" in {
-      def substitute(rel: String, newRef: String)(html: String) = {
-        val (pattern, newValue) = KifiSiteRouter.substituteLink(rel, newRef)
-        pattern.replaceAllIn(html, newValue)
-      }
-      substitute("canonical", "http://www.kifi.com")("""<link rel="canonical" href="http://www.lemonde.fr" />""") === """<link rel="canonical" href="http://www.kifi.com"/>"""
-    }
-
   }
+
 }
