+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/website/MarketingSiteRouterTest.scala
@@ -52,6 +52,26 @@ class MarketingSiteRouterTest extends Specification with ShoeboxApplicationInjec
         landing === "index.3"
       }
     }
+
+    "substitute meta properties" in {
+      def substitute(property: String, newContent: String)(html: String) = {
+        val (pattern, newValue) = MarketingSiteRouter.substituteMetaProperty(property, newContent)
+        pattern.replaceAllIn(html, newValue)
+      }
+      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge" />""") === """<meta property="og:title" content="Hi there!"/>"""
+      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge">""") === """<meta property="og:title" content="Hi there!"/>"""
+      substitute("og:title", "Hi there!")("""<meta property="og:title" content="Connecting People With Knowledge" >""") === """<meta property="og:title" content="Hi there!"/>"""
+      substitute("og:description", "Hi there!")("""<meta property="og:description" content="Connecting People With Knowledge" />""") === """<meta property="og:description" content="Hi there!"/>"""
+      substitute("og:title", "Hi there!")("""<meta property="og:description" content="Connecting People With Knowledge" />""") === """<meta property="og:description" content="Connecting People With Knowledge" />"""
+    }
+
+    "substitute link tags" in {
+      def substitute(rel: String, newRef: String)(html: String) = {
+        val (pattern, newValue) = MarketingSiteRouter.substituteLink(rel, newRef)
+        pattern.replaceAllIn(html, newValue)
+      }
+      substitute("canonical", "http://www.kifi.com")("""<link rel="canonical" href="http://www.lemonde.fr" />""") === """<link rel="canonical" href="http://www.kifi.com"/>"""
+    }
   }
 
 }
